New upstream version 10.5.7

+++ storage/columnstore/columnstore/storage-manager/src/S3Storage.cpp
@@ -24,6 +24,9 @@
 #include <sys/types.h>
 #include <boost/filesystem.hpp>
 #include <iostream>
+#include <boost/uuid/uuid.hpp>
+#include <boost/uuid/uuid_io.hpp>
+#include <boost/uuid/random_generator.hpp>
 #include "Utilities.h"
 
 using namespace std;
@@ -38,7 +41,8 @@ inline bool retryable_error(uint8_t s3err)
         s3err == MS3_ERR_REQUEST_ERROR ||
         s3err == MS3_ERR_OOM ||
         s3err == MS3_ERR_IMPOSSIBLE ||
-        s3err == MS3_ERR_SERVER
+        s3err == MS3_ERR_SERVER ||
+        s3err == MS3_ERR_AUTH_ROLE
     );
 }
 
@@ -86,7 +90,7 @@ S3Storage::ScopedConnection::~ScopedConnection()
     s3->returnConnection(conn);
 }
 
-S3Storage::S3Storage()
+S3Storage::S3Storage(bool skipRetry) : skipRetryableErrors(skipRetry)
 {
     /*  Check creds from envvars
         Get necessary vars from config
@@ -98,6 +102,9 @@ S3Storage::S3Storage()
         " or setting aws_access_key_id and aws_secret_access_key in storagemanager.cnf";
     key = config->getValue("S3", "aws_access_key_id");
     secret = config->getValue("S3", "aws_secret_access_key");
+    IAMrole = config->getValue("S3", "iam_role_name");
+    STSendpoint = config->getValue("S3", "sts_endpoint");
+    STSregion = config->getValue("S3", "sts_region");
     if (key.empty())
     {
         char *_key_id = getenv("AWS_ACCESS_KEY_ID");
@@ -118,7 +125,7 @@ S3Storage::S3Storage()
         }
         secret = _secret_id;
     }
-    
+
     region = config->getValue("S3", "region");
     bucket = config->getValue("S3", "bucket");
     prefix = config->getValue("S3", "prefix");
@@ -132,7 +139,7 @@ S3Storage::S3Storage()
     endpoint = config->getValue("S3", "endpoint");
 
     ms3_library_init();
-    //ms3_debug(true);
+    //ms3_debug();
     testConnectivityAndPerms();
 }
 
@@ -154,7 +161,11 @@ void S3Storage::testConnectivityAndPerms()
 {
     boost::shared_array<uint8_t> testObj(new uint8_t[1]);
     testObj[0] = 0;
-    string testObjKey("connectivity_test");
+    boost::uuids::uuid u = boost::uuids::random_generator()();
+    ostringstream oss;
+    oss << u << "connectivity_test";
+    
+    string testObjKey = oss.str();
     
     int err = putObject(testObj, 1, testObjKey);
     if (err)
@@ -220,11 +231,17 @@ int S3Storage::getObject(const string &_sourceKey, boost::shared_array<uint8_t> 
     string sourceKey = prefix + _sourceKey;
     
     ms3_st *creds = getConnection();
+    if (!creds)
+    {
+        logger->log(LOG_ERR, "S3Storage::getObject(): failed to GET, S3Storage::getConnection() returned NULL on init");
+        errno = EINVAL;
+        return -1;
+    }
     ScopedConnection sc(this, creds);
     
     do {
         err = ms3_get(creds, bucket.c_str(), sourceKey.c_str(), &_data, &len);
-        if (err && retryable_error(err))
+        if (err && (!skipRetryableErrors && retryable_error(err)))
         { 
             if (ms3_server_error(creds))
                 logger->log(LOG_ERR, "S3Storage::getObject(): failed to GET, server says '%s'.  bucket = %s, key = %s."
@@ -232,9 +249,13 @@ int S3Storage::getObject(const string &_sourceKey, boost::shared_array<uint8_t> 
             else 
                 logger->log(LOG_ERR, "S3Storage::getObject(): failed to GET, got '%s'.  bucket = %s, key = %s.  Retrying...",
                     s3err_msgs[err], bucket.c_str(), sourceKey.c_str());
+            if(!IAMrole.empty())
+            {
+                ms3_assume_role(creds);
+            }
             sleep(5);
         }
-    } while (err && retryable_error(err));
+    } while (err && (!skipRetryableErrors && retryable_error(err)));
     if (err)
     {
         if (ms3_server_error(creds))
@@ -310,11 +331,17 @@ int S3Storage::putObject(const boost::shared_array<uint8_t> data, size_t len, co
     string destKey = prefix + _destKey;
     uint8_t s3err;
     ms3_st *creds = getConnection();
+    if (!creds)
+    {
+        logger->log(LOG_ERR, "S3Storage::putObject(): failed to PUT, S3Storage::getConnection() returned NULL on init");
+        errno = EINVAL;
+        return -1;
+    }
     ScopedConnection sc(this, creds);
     
     do {
         s3err = ms3_put(creds, bucket.c_str(), destKey.c_str(), data.get(), len);
-        if (s3err && retryable_error(s3err))
+        if (s3err && (!skipRetryableErrors && retryable_error(s3err)))
         {
             if (ms3_server_error(creds))
                 logger->log(LOG_ERR, "S3Storage::putObject(): failed to PUT, server says '%s'.  bucket = %s, key = %s."
@@ -322,9 +349,13 @@ int S3Storage::putObject(const boost::shared_array<uint8_t> data, size_t len, co
             else
                 logger->log(LOG_ERR, "S3Storage::putObject(): failed to PUT, got '%s'.  bucket = %s, key = %s."
                     "  Retrying...", s3err_msgs[s3err], bucket.c_str(), destKey.c_str());
+            if(!IAMrole.empty())
+            {
+                ms3_assume_role(creds);
+            }
             sleep(5);
         }
-    } while (s3err && retryable_error(s3err));
+    } while (s3err && (!skipRetryableErrors && retryable_error(s3err)));
     if (s3err)
     {
         if (ms3_server_error(creds))
@@ -344,11 +375,17 @@ int S3Storage::deleteObject(const string &_key)
     uint8_t s3err;
     string key = prefix + _key;
     ms3_st *creds = getConnection();
+    if (!creds)
+    {
+        logger->log(LOG_ERR, "S3Storage::deleteObject(): failed to DELETE, S3Storage::getConnection() returned NULL on init");
+        errno = EINVAL;
+        return -1;
+    }
     ScopedConnection sc(this, creds);
-        
+
     do {
         s3err = ms3_delete(creds, bucket.c_str(), key.c_str());
-        if (s3err && s3err != MS3_ERR_NOT_FOUND && retryable_error(s3err))
+        if (s3err && s3err != MS3_ERR_NOT_FOUND && (!skipRetryableErrors && retryable_error(s3err)))
         {
             if (ms3_server_error(creds))
                 logger->log(LOG_ERR, "S3Storage::deleteObject(): failed to DELETE, server says '%s'.  bucket = %s, key = %s."
@@ -356,9 +393,13 @@ int S3Storage::deleteObject(const string &_key)
             else
                 logger->log(LOG_ERR, "S3Storage::deleteObject(): failed to DELETE, got '%s'.  bucket = %s, key = %s.  Retrying...",
                     s3err_msgs[s3err], bucket.c_str(), key.c_str());
+            if(!IAMrole.empty())
+            {
+                ms3_assume_role(creds);
+            }
             sleep(5);
         }
-    } while (s3err && s3err != MS3_ERR_NOT_FOUND && retryable_error(s3err));
+    } while (s3err && s3err != MS3_ERR_NOT_FOUND && (!skipRetryableErrors && retryable_error(s3err)));
     
     if (s3err != 0 && s3err != MS3_ERR_NOT_FOUND)
     {
@@ -378,12 +419,18 @@ int S3Storage::copyObject(const string &_sourceKey, const string &_destKey)
     string sourceKey = prefix + _sourceKey, destKey = prefix + _destKey;
     uint8_t s3err;
     ms3_st *creds = getConnection();
+    if (!creds)
+    {
+        logger->log(LOG_ERR, "S3Storage::copyObject(): failed to copy, S3Storage::getConnection() returned NULL on init");
+        errno = EINVAL;
+        return -1;
+    }
     ScopedConnection sc(this, creds);
     
     do 
     {
         s3err = ms3_copy(creds, bucket.c_str(), sourceKey.c_str(), bucket.c_str(), destKey.c_str());
-        if (s3err && retryable_error(s3err))
+        if (s3err && (!skipRetryableErrors && retryable_error(s3err)))
         {
             if (ms3_server_error(creds))
                 logger->log(LOG_ERR, "S3Storage::copyObject(): failed to copy, server says '%s'.  bucket = %s, srckey = %s, "
@@ -391,9 +438,13 @@ int S3Storage::copyObject(const string &_sourceKey, const string &_destKey)
             else
                 logger->log(LOG_ERR, "S3Storage::copyObject(): failed to copy, got '%s'.  bucket = %s, srckey = %s, "
                     " destkey = %s.  Retrying...", s3err_msgs[s3err], bucket.c_str(), sourceKey.c_str(), destKey.c_str());
+            if(!IAMrole.empty())
+            {
+                ms3_assume_role(creds);
+            }
             sleep(5);
         }
-    } while (s3err && retryable_error(s3err));
+    } while (s3err && (!skipRetryableErrors && retryable_error(s3err)));
     
     if (s3err) 
     {
@@ -428,11 +479,17 @@ int S3Storage::exists(const string &_key, bool *out)
     uint8_t s3err;
     ms3_status_st status;
     ms3_st *creds = getConnection();
+    if (!creds)
+    {
+        logger->log(LOG_ERR, "S3Storage::exists(): failed to HEAD, S3Storage::getConnection() returned NULL on init");
+        errno = EINVAL;
+        return -1;
+    }
     ScopedConnection sc(this, creds);
     
     do {
         s3err = ms3_status(creds, bucket.c_str(), key.c_str(), &status);
-        if (s3err && s3err != MS3_ERR_NOT_FOUND && retryable_error(s3err))
+        if (s3err && s3err != MS3_ERR_NOT_FOUND && (!skipRetryableErrors && retryable_error(s3err)))
         {
             if (ms3_server_error(creds))
                 logger->log(LOG_ERR, "S3Storage::exists(): failed to HEAD, server says '%s'.  bucket = %s, key = %s."
@@ -440,9 +497,13 @@ int S3Storage::exists(const string &_key, bool *out)
             else
                 logger->log(LOG_ERR, "S3Storage::exists(): failed to HEAD, got '%s'.  bucket = %s, key = %s.  Retrying...",
                     s3err_msgs[s3err], bucket.c_str(), key.c_str());
+            if(!IAMrole.empty())
+            {
+                ms3_assume_role(creds);
+            }
             sleep(5);
         }
-    } while (s3err && s3err != MS3_ERR_NOT_FOUND && retryable_error(s3err));
+    } while (s3err && s3err != MS3_ERR_NOT_FOUND && (!skipRetryableErrors && retryable_error(s3err)));
     
     if (s3err != 0 && s3err != MS3_ERR_NOT_FOUND)
     {
@@ -483,11 +544,27 @@ ms3_st * S3Storage::getConnection()
             
     // get a connection
     ms3_st *ret = NULL;
+    uint8_t res = 0;
     if (freeConns.empty())
     {
         ret = ms3_init(key.c_str(), secret.c_str(), region.c_str(), (endpoint.empty() ? NULL : endpoint.c_str()));
         if (ret == NULL)
             logger->log(LOG_ERR, "S3Storage::getConnection(): ms3_init returned NULL, no specific info to report");
+        if(!IAMrole.empty())
+        {
+            res = ms3_init_assume_role(ret, (IAMrole.empty() ? NULL : IAMrole.c_str()),
+                                            (STSendpoint.empty() ? NULL : STSendpoint.c_str()),
+                                            (STSregion.empty() ? NULL : STSregion.c_str()));
+            if (res)
+            {
+                // Something is wrong with the assume role so abort as if the ms3_init failed
+                logger->log(LOG_ERR, "S3Storage::getConnection(): ERROR: ms3_init_assume_role. Verify iam_role_name = %s, aws_access_key_id, aws_secret_access_key values. Also check sts_region and sts_endpoint if configured.",IAMrole.c_str());
+                if (ms3_server_error(ret))
+                    logger->log(LOG_ERR, "S3Storage::getConnection(): ms3_error: server says '%s'  role name = %s", ms3_server_error(ret), IAMrole.c_str());
+                ms3_deinit(ret);
+                ret = NULL;
+            }
+        }
         //assert(connMutexes[ret].try_lock());
         s.unlock();
         return ret;
